Initial split of two algorithms
Re #13282
